requirements.txt
django_heroku
djangorestframework
psycopg2-binary
requests